refactor: 분수 타입 api type 으로 이동
